@@ -20,10 +20,12 @@
         "resolve-url-loader": "^2.3.1",
         "sass": "^1.20.1",
         "sass-loader": "7.*",
-        "vue": "^2.5.17",
+        "vue": "^2.6.10",
         "vue-template-compiler": "^2.6.10"
     },
     "dependencies": {
-        "toastr": "^2.1.4"
+        "toastr": "^2.1.4",
+        "vue-router": "^3.1.3",
+        "vuetify": "^2.1.10"
     }
 }
