@@ -1,33 +1,22 @@
-/**
- * First we will load all of this project's JavaScript dependencies which
- * includes Vue and other libraries. It is a great starting point when
- * building robust, powerful web applications using Vue and Laravel.
- */
+import './bootstrap';
+import Vue from 'vue';
+import Vuetify from 'vuetify';
 
-require('./bootstrap');
+//routes
+import Routes from '@/js/routes.js';
 
-window.Vue = require('vue');
+//views
+import App from '@/js/views/App';
 
-/**
- * The following block of code may be used to automatically register your
- * Vue components. It will recursively scan this directory for the Vue
- * components and automatically register them with their "basename".
- *
- * Eg. ./components/ExampleComponent.vue -> <example-component></example-component>
- */
-
-// const files = require.context('./', true, /\.vue$/i)
-// files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default))
-
-Vue.component('example-component', require('./components/ExampleComponent.vue').default);
 Vue.component('auth', require('./components/guest/Auth.vue').default);
 
-/**
- * Next, we will create a fresh Vue application instance and attach it to
- * the page. Then, you may begin adding components to this application
- * or customize the JavaScript scaffolding to fit your unique needs.
- */
+Vue.use(Vuetify);
 
 const app = new Vue({
     el: '#app',
+    vuetify : new Vuetify(),
+    router: Routes,
+    render: h => h(App),
 });
+
+export default app;
