@@ -5,8 +5,8 @@ namespace App\Http\Controllers;
 
 class IndexController extends Controller
 {
-    public function index(){
-        return view('home');
-        //return 'ok';
+    public function index()
+    {
+        return view('layouts.app');
     }
 }
